Seperated the collector out of the shooter class; rewrote the commands for the new collector and shooter subsystem
