+++ src/org/usfirst/frc/team5590/robot/commands/Collect.java
@@ -11,15 +11,15 @@ import org.usfirst.frc.team5590.robot.subsystems.*;
 public class Collect extends Command {
 	
     public Collect() {
-    	requires(Robot.shooter);
+    	requires(Robot.collector);
     	// TODO: Timeout change based on tests
     	setTimeout(1.0);
     }
 
     protected void initialize() {
     	// TODO: Speed changed according to tests
-    	Robot.shooter.stopCollector();
-    	Robot.shooter.setCollectorSpeed(.0);
+    	Robot.collector.stopCollector();
+    	Robot.collector.setCollectorSpeed(.0);
     }
 
     protected void execute() {
@@ -32,12 +32,12 @@ public class Collect extends Command {
 
     // Called once after isFinished returns true
     protected void end() {
-    	Robot.shooter.stopCollector();
+    	Robot.collector.stopCollector();
     }
 
     // Called when another command which requires one or more of the same
     // subsystems is scheduled to run
     protected void interrupted() {
-    	Robot.shooter.stopCollector();
+    	Robot.collector.stopCollector();
     }
 }
